Modifications in Headline Remover

+++ Headline Remover/ThresholedValue.cpp
@@ -114,7 +114,7 @@ float ThresholedValue::GetThresholed()
 				count++;
 			}
 		}//end of for
-		float idx = sumI/count;
+		float idx = (float)sumI/count;
 		level = (idx - 1)/(num_bins-1);
 	}// end of if
 	else
